# hms-project/docs/ambulance-management-openapi.yaml
openapi: 3.0.3
info:
  title: HMS Ambulance Management API
  description: >-
    This OpenAPI specification documents all ambulance management related
    endpoints, including ambulance inventory management, dispatch system,
    real-time location tracking, driver management, service history,
    and billing integration.
  version: 1.0.0
servers:
  - url: https://hms.example.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Local dev server
...
paths:
  /ambulances/inventory:
    get:
      tags: [Ambulance Inventory]
      summary: Get ambulance inventory with optional filtering
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [AVAILABLE, IN_USE, MAINTENANCE, OUT_OF_SERVICE]
        - in: query
          name: vehicleType
          schema:
            type: string
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: List of ambulances with pagination metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AmbulanceListResponse'
    post:
      tags: [Ambulance Inventory]
      summary: Add a new ambulance to inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmbulanceInput'
      responses:
        '201':
          description: Ambulance created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ambulance'
    put:
      tags: [Ambulance Inventory]
      summary: Update ambulance details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/AmbulanceInput'
                - type: object
                  properties:
                    id:
                      type: string
                  required: [id]
      responses:
        '200':
          description: Ambulance updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ambulance'
    delete:
      tags: [Ambulance Inventory]
      summary: Delete an ambulance
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Ambulance deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
components:
  schemas:
    PagedResponse:
      type: object
      properties:
        data:
          type: array
          items: { type: object }
        meta:
          type: object
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
    AmbulanceListResponse:
      allOf:
        - $ref: '#/components/schemas/PagedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Ambulance'
    AmbulanceInput:
      type: object
      required: [registrationNumber, vehicleModel, vehicleType, manufacturingYear, capacity]
      properties:
        registrationNumber:
          type: string
        vehicleModel:
          type: string
        vehicleType:
          type: string
          enum: [BASIC_LIFE_SUPPORT, ADVANCED_LIFE_SUPPORT, PATIENT_TRANSPORT, NEONATAL, MOBILE_ICU]
        manufacturingYear:
          type: integer
        capacity:
          type: integer
        features:
          type: array
          items:
            type: string
        purchaseDate:
          type: string
          format: date
        insuranceExpiry:
          type: string
          format: date
    Ambulance:
      allOf:
        - $ref: '#/components/schemas/AmbulanceInput'
        - type: object
          properties:
            id:
              type: string
            status:
              type: string
              enum: [AVAILABLE, IN_USE, MAINTENANCE, OUT_OF_SERVICE]
            lastMaintenance:
              type: string
              format: date-time
            nextMaintenance:
              type: string
              format: date-time
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...